feat(services/gcs): implement write returns metadata
