+++ core/src/services/gcs/backend.rs
@@ -15,12 +15,10 @@
 // specific language governing permissions and limitations
 // under the License.
 
-use std::collections::HashMap;
 use std::fmt::Debug;
 use std::fmt::Formatter;
 use std::sync::Arc;
 
-use bytes::Buf;
 use http::Response;
 use http::StatusCode;
 use log::debug;
@@ -28,8 +26,6 @@ use reqsign::GoogleCredentialLoader;
 use reqsign::GoogleSigner;
 use reqsign::GoogleTokenLoad;
 use reqsign::GoogleTokenLoader;
-use serde::Deserialize;
-use serde_json;
 
 use super::core::*;
 use super::delete::GcsDeleter;
@@ -433,37 +429,7 @@ impl Access for GcsBackend {
         }
 
         let slc = resp.into_body();
-
-        let meta: GetObjectJsonResponse =
-            serde_json::from_reader(slc.reader()).map_err(new_json_deserialize_error)?;
-
-        let mut m = Metadata::new(EntryMode::FILE);
-
-        m.set_etag(&meta.etag);
-        m.set_content_md5(&meta.md5_hash);
-
-        let size = meta
-            .size
-            .parse::<u64>()
-            .map_err(|e| Error::new(ErrorKind::Unexpected, "parse u64").set_source(e))?;
-        m.set_content_length(size);
-        if !meta.content_type.is_empty() {
-            m.set_content_type(&meta.content_type);
-        }
-
-        if !meta.content_encoding.is_empty() {
-            m.set_content_encoding(&meta.content_encoding);
-        }
-
-        if !meta.cache_control.is_empty() {
-            m.set_cache_control(&meta.cache_control);
-        }
-
-        m.set_last_modified(parse_datetime_from_rfc3339(&meta.updated)?);
-
-        if !meta.metadata.is_empty() {
-            m.with_user_metadata(meta.metadata);
-        }
+        let m = GcsCore::get_metadata_from_response(path, slc)?;
 
         Ok(RpStat::new(m))
     }
@@ -547,89 +513,5 @@ impl Access for GcsBackend {
             parts.uri,
             parts.headers,
         )))
-    }
-}
-
-/// The raw json response returned by [`get`](https://cloud.google.com/storage/docs/json_api/v1/objects/get)
-#[derive(Debug, Default, Deserialize)]
-#[serde(default, rename_all = "camelCase")]
-struct GetObjectJsonResponse {
-    /// GCS will return size in string.
-    ///
-    /// For example: `"size": "56535"`
-    size: String,
-    /// etag is not quoted.
-    ///
-    /// For example: `"etag": "CKWasoTgyPkCEAE="`
-    etag: String,
-    /// RFC3339 styled datetime string.
-    ///
-    /// For example: `"updated": "2022-08-15T11:33:34.866Z"`
-    updated: String,
-    /// Content md5 hash
-    ///
-    /// For example: `"md5Hash": "fHcEH1vPwA6eTPqxuasXcg=="`
-    md5_hash: String,
-    /// Content type of this object.
-    ///
-    /// For example: `"contentType": "image/png",`
-    content_type: String,
-    /// Content encoding of this object
-    ///
-    /// For example: "contentEncoding": "br"
-    content_encoding: String,
-    /// Cache-Control directive for the object data.
-    cache_control: String,
-    /// Custom metadata of this object.
-    ///
-    /// For example: `"metadata" : { "my-key": "my-value" }`
-    metadata: HashMap<String, String>,
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn test_deserialize_get_object_json_response() {
-        let content = r#"{
-  "kind": "storage#object",
-  "id": "example/1.png/1660563214863653",
-  "selfLink": "https://www.googleapis.com/storage/v1/b/example/o/1.png",
-  "mediaLink": "https://content-storage.googleapis.com/download/storage/v1/b/example/o/1.png?generation=1660563214863653&alt=media",
-  "name": "1.png",
-  "bucket": "example",
-  "generation": "1660563214863653",
-  "metageneration": "1",
-  "contentType": "image/png",
-  "contentEncoding": "br",
-  "cacheControl": "public, max-age=3600",
-  "storageClass": "STANDARD",
-  "size": "56535",
-  "md5Hash": "fHcEH1vPwA6eTPqxuasXcg==",
-  "crc32c": "j/un9g==",
-  "etag": "CKWasoTgyPkCEAE=",
-  "timeCreated": "2022-08-15T11:33:34.866Z",
-  "updated": "2022-08-15T11:33:34.866Z",
-  "timeStorageClassUpdated": "2022-08-15T11:33:34.866Z",
-  "metadata" : {
-    "location" : "everywhere"
-  }
-}"#;
-
-        let meta: GetObjectJsonResponse =
-            serde_json::from_str(content).expect("json Deserialize must succeed");
-
-        assert_eq!(meta.size, "56535");
-        assert_eq!(meta.updated, "2022-08-15T11:33:34.866Z");
-        assert_eq!(meta.md5_hash, "fHcEH1vPwA6eTPqxuasXcg==");
-        assert_eq!(meta.etag, "CKWasoTgyPkCEAE=");
-        assert_eq!(meta.content_type, "image/png");
-        assert_eq!(meta.content_encoding, "br".to_string());
-        assert_eq!(meta.cache_control, "public, max-age=3600".to_string());
-        assert_eq!(
-            meta.metadata,
-            HashMap::from_iter([("location".to_string(), "everywhere".to_string())])
-        );
     }
 }
