+++ core/src/services/gcs/writer.rs
@@ -62,7 +62,10 @@ impl oio::MultipartWrite for GcsWriter {
         let status = resp.status();
 
         match status {
-            StatusCode::CREATED | StatusCode::OK => Ok(Metadata::default()),
+            StatusCode::CREATED | StatusCode::OK => {
+                let metadata = GcsCore::get_metadata_from_response(&self.path, resp.into_body())?;
+                Ok(metadata)
+            }
             _ => Err(parse_error(resp)),
         }
     }
@@ -139,7 +142,8 @@ impl oio::MultipartWrite for GcsWriter {
         if !resp.status().is_success() {
             return Err(parse_error(resp));
         }
-        Ok(Metadata::default())
+        let metadata = GcsCore::get_metadata_from_response(&self.path, resp.into_body())?;
+        Ok(metadata)
     }
 
     async fn abort_part(&self, upload_id: &str) -> Result<()> {
